+++ src/ska_mid_cbf_fhs_vcc/wideband_frequency_shifter/wideband_frequency_shifter_manager.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from typing import Optional
 
 from dataclasses_json import DataClassJsonMixin
 from ska_mid_cbf_fhs_common import BaseIPBlockManager
@@ -8,6 +9,7 @@ from ska_mid_cbf_fhs_vcc.wideband_frequency_shifter.wideband_frequency_shifter_s
 
 @dataclass
 class WidebandFrequencyShifterConfig(DataClassJsonMixin):
+    transaction_id: Optional[str] = None
     shift_frequency: float = 0.0